[Bronze V] Title: 이상한 기호, Time: 48 ms, Memory: 5484 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cs
@@ -1,2 +1,2 @@
-int[] arr = Array.ConvertAll(Console.ReadLine().Split(' '),int.Parse);
+long[] arr = Array.ConvertAll(Console.ReadLine().Split(' '), long.Parse);
 Console.WriteLine((arr[0] + arr[1]) * (arr[0] - arr[1]));
